convert float to double

+++ moveit_core/robot_model/src/floating_joint_model.cpp
@@ -173,7 +173,7 @@ bool FloatingJointModel::satisfiesPositionBounds(const double* values, const Bou
   if (values[2] < bounds[2].min_position_ - margin || values[2] > bounds[2].max_position_ + margin)
     return false;
   double norm_sqr = values[3] * values[3] + values[4] * values[4] + values[5] * values[5] + values[6] * values[6];
-  return fabs(norm_sqr - 1.0) <= std::numeric_limits<float>::epsilon() * 10.0;
+  return fabs(norm_sqr - 1.0) <= std::numeric_limits<double>::epsilon() * 10.0;
 }
 
 bool FloatingJointModel::normalizeRotation(double* values) const
