+++ moveit_ros/visualization/rviz_plugin_render_tools/include/moveit/rviz_plugin_render_tools/render_shapes.h
@@ -59,8 +59,8 @@ public:
 
   void renderShape(Ogre::SceneNode* node, const shapes::Shape* s, const Eigen::Isometry3d& p,
                    OctreeVoxelRenderMode octree_voxel_rendering, OctreeVoxelColorMode octree_color_mode,
-                   const Ogre::ColourValue& color, float alpha);
-  void updateShapeColors(float r, float g, float b, float a);
+                   const Ogre::ColourValue& color, double alpha);
+  void updateShapeColors(double r, double g, double b, double a);
   void clear();
 
 private:
